ludora_backend/requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings

# Security / Authentication
passlib[bcrypt]
python-jose[cryptography]
bcrypt

# Database (Tortoise ORM)
tortoise-orm
asyncpg # For PostgreSQL, if you choose to use it. Or mysqlclient for MySQL, etc.
aerich

# Question Generation
mathgenerator

# Rate Limiting
slowapi

# Testing
pytest
pytest-asyncio
httpx  # For async HTTP requests to the test client
# faker # Optional: for generating fake data in tests

# AI Model Inference
onnxruntime
transformers
sentencepiece
